refactor: websocket classes are now inside their own folder
